% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtlr_cv.R
\name{mtlr_cv}
\alias{mtlr_cv}
\title{Primary MTLR function. Run the mode}
\usage{
mtlr_cv(formula, data, nintervals = NULL, normalize = T,
  C1 = c(0.001, 0.01, 0.1, 1, 10, 100, 1000), train_biases = T,
  loss = c("ll"), nfolds = 5, foldtype = c("fullstrat",
  "censorstrat", "random"), threshold = 1e-05, maxit = 5000,
  lower = -20, upper = 20)
}
\arguments{
\item{formula}{Surv formula for survival data.}

\item{data}{Surv data.}

\item{nintervals}{Number of time intervals.}

\item{normalize}{Boolean specifying if normalization should occur. Strongly suggested.}

\item{C1}{a vector of regularization parameters to test. All values must be non-negative.}

\item{train_biases}{Boolean specifying whether to run an interation training biases first.}

\item{loss}{a string indicating the loss to optimize for which to choose the regularization parameter. Currently on the log-likelihood (ll) is supported.}

\item{nfolds}{the number of internal cross validation folds}

\item{foldtype}{type of cross validation folds}

\item{threshold}{The threshold for optim}

\item{maxit}{The maximum interations for optim}

\item{lower}{The lower L-BFGS-B optim}

\item{upper}{The upper.}
}
\description{
Primary MTLR function. Run the mode
}
