% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtlr_objVal}
\alias{mtlr_objVal}
\title{Calculate objective value.}
\usage{
mtlr_objVal(params, yval, featureValue, C1, delta)
}
\arguments{
\item{params}{A vector of starting weights b1,b2,...,bm,theta11,theta12,...theta1m, ...theta p 1,...,theta p m}

\item{yval}{A matrix of indicators of death}

\item{featureValue}{A matrix of feature values.}

\item{C1}{The regularization term.}

\item{delta}{The event status indicator (1= death, 0 = censored), sorted.}
}
\description{
Calculate objective value.
}
