% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mtlr_predict}
\alias{mtlr_predict}
\title{Predict survival curves.}
\usage{
mtlr_predict(params, featureValue)
}
\arguments{
\item{params}{A vector of starting weights b1,b2,...,bm,theta11,theta12,...theta1m, ...theta p 1,...,theta p m}

\item{featureValue}{A matrix of feature values.}
}
\description{
Predict survival curves.
}
