% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{plot.mtlr}
\alias{plot.mtlr}
\title{Plot the weights of the mtlr object across time. The numfeatures = 5 default plots the the 5 most influential features
as determined by the sum of absolute values of feature weights across time. The bias is ignored in these calculations.}
\usage{
\method{plot}{mtlr}(x, numfeatures = 5, weightnames = c(), digits, ...)
}
\arguments{
\item{x}{an object of class mtlr, returned by the mtlr function.}

\item{numfeatures}{the number of weight to plot. Default is 5.}

\item{weightnames}{the names of the specific weight to plot. These should correspond to the names
in mtlr$weight_matrix. If weightnames are supplied, then numfeatures is ignored.}

\item{digits}{the number of digits to round to for the time points.}

\item{...}{for future methods}
}
\description{
Plot the weights of the mtlr object across time. The numfeatures = 5 default plots the the 5 most influential features
as determined by the sum of absolute values of feature weights across time. The bias is ignored in these calculations.
}
