% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcurves.R
\name{plotcurves}
\alias{plotcurves}
\title{Plot the survival curves returned from MTLR with a spline fit. If over 15 survival curves are plotted the legend is removed.}
\usage{
plotcurves(curves, index = 1, color = c(), xlim = c())
}
\arguments{
\item{curves}{survival curves formatted the same as those from predict.mtlr. The time points must be in the first column.}

\item{index}{the index of the observation to plot. Here "1" will refer to the second column of curves}

\item{color}{the color of the plotted survival curve. The length of color must match the length of index.}

\item{xlim}{the limits of the x-axis (must be a 2 length vector).}
}
\description{
Plot the survival curves returned from MTLR with a spline fit. If over 15 survival curves are plotted the legend is removed.
}
