% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{predict.mtlr}
\alias{predict.mtlr}
\title{Predict method for MTLR class.}
\usage{
\method{predict}{mtlr}(object, newdata, type = c("survivalcurve",
  "prob_event", "mean_time", "median_time"), add_zero = T, ...)
}
\arguments{
\item{object}{an object with class mtlr, generated by mtlr().}

\item{newdata}{a dataframe consisting of the same variables mtlr was trained on.}

\item{type}{the type of prediciton. Currently only response is supported (response gives survival curves).}

\item{add_zero}{a boolean specifying}

\item{...}{for future methods.}
}
\description{
Predict method for MTLR class.
}
