% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.mtlr}
\alias{print.mtlr}
\title{prints an mtlr object}
\usage{
\method{print}{mtlr}(x, digits = max(3, getOption("digits")), ...)
}
\arguments{
\item{x}{an mtlr object}

\item{digits}{The number of digits to print mtlr weights}

\item{...}{for future methods.}
}
\description{
prints an mtlr object
}
