% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtlr.R
\name{mtlr}
\alias{mtlr}
\title{Primary MTLR function. Run the mode}
\usage{
mtlr(formula, data, nintervals = NULL, normalize = T, C1 = 1,
  train_biases = T, threshold = 1e-05, maxit = 5000, lower = -20,
  upper = 20)
}
\arguments{
\item{formula}{Surv formula for survival data.}

\item{data}{Surv data.}

\item{nintervals}{Number of time intervals.}

\item{normalize}{Boolean specifying if normalization should occur. Strongly suggested.}

\item{C1}{The regularization parameter}

\item{train_biases}{Boolean specifying whether to run an interation training biases first.}

\item{threshold}{The threshold for optim}

\item{maxit}{The maximum interations for optim}

\item{lower}{The lower L-BFGS-B optim}

\item{upper}{The upper.}
}
\description{
Primary MTLR function. Run the mode
}
